Restructure logging classes
